docs(example): update example code
- add react-native-safe-area-context, vector-ionc setting
- update react-native, expo version
- add close button
