+++ example/package.json
@@ -10,14 +10,16 @@
   },
   "dependencies": {
     "@expo/metro-runtime": "~5.0.4",
-    "expo": "~53.0.15",
+    "@expo/vector-icons": "^14.1.0",
+    "expo": "~53.0.16",
     "expo-status-bar": "~2.2.3",
     "react": "19.0.0",
     "react-dom": "19.0.0",
-    "react-native": "0.79.4",
+    "react-native": "0.79.5",
     "react-native-gesture-handler": "~2.24.0",
     "react-native-gesture-image-viewer": "workspace:*",
     "react-native-reanimated": "~3.17.4",
+    "react-native-safe-area-context": "~5.4.0",
     "react-native-web": "~0.20.0"
   },
   "devDependencies": {
