+++ package.json
@@ -64,7 +64,7 @@
     "@changesets/cli": "^2.29.5",
     "@commitlint/config-conventional": "^19.6.0",
     "@evilmartians/lefthook": "^1.5.0",
-    "@react-native/babel-preset": "0.78.2",
+    "@react-native/babel-preset": "0.79.5",
     "@types/babel__core": "^7",
     "@types/jest": "^29.5.5",
     "@types/react": "^19.0.12",
@@ -72,7 +72,7 @@
     "del-cli": "^5.1.0",
     "jest": "^29.7.0",
     "react": "19.0.0",
-    "react-native": "0.79.4",
+    "react-native": "0.79.5",
     "react-native-builder-bob": "^0.40.12",
     "react-native-gesture-handler": "^2.24.0",
     "react-native-reanimated": "^3.18.0",
